Updated scoring/promotion for pawns; Changed move execution at Pieces level; Updated Queen - should be working; Added Params object to Game file - this object stores static global values. Next: build I/O layer for unit testing using game states read from files.

+++ src/State.scala
@@ -7,7 +7,7 @@
   * Note that its "value" field is the only one that is mutable - this is to facilitate
   * ttable functionality.
   * */
-class State(val on_move: Player, var b_value: Float = 0, var w_value: Float = 0, val pieces: List[Piece]){
+class State(val on_move: Player, var b_value: Double = 0, var w_value: Double = 0, val pieces: List[Piece]){
   //XXX: Potentially add other fields in support of heuristic evaluation
 
   override def toString: String = {
